# leadvirt_codex_docs_package_v2/templates/openapi_outline.yaml
openapi: 3.0.3
info:
  title: LeadVirt.ai API
  version: 0.1.0
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /leads:
    get:
      summary: List leads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lead list
    post:
      summary: Create lead
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Lead created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT